src/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.5
django-cors-headers==3.6.0
django-heroku==0.3.1
django-nested-admin==3.3.3
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.4
python-monkey-business==1.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
yapf==0.30.0
